+++ collections/mod.ts
@@ -43,3 +43,6 @@ export * from "./drop_last_while.ts";
 export * from "./reduce_groups.ts";
 export * from "./sample.ts";
 export * from "./running_reduce.ts";
+export * from "./binary_heap.ts";
+export * from "./bs_tree.ts";
+export * from "./rb_tree.ts";