+++ node/crypto.ts
@@ -7,6 +7,12 @@ import { pbkdf2, pbkdf2Sync } from "./_crypto/pbkdf2.ts";
 import { scrypt, scryptSync } from "./_crypto/scrypt.ts";
 import { timingSafeEqual } from "./_crypto/timingSafeEqual.ts";
 import { createHash, getHashes, Hash } from "./_crypto/hash.ts";
+import {
+  privateDecrypt,
+  privateEncrypt,
+  publicDecrypt,
+  publicEncrypt,
+} from "./_crypto/crypto_browserify/public_encrypt/mod.js";
 
 const randomUUID = () => crypto.randomUUID();
 const webcrypto = crypto;
@@ -20,6 +26,10 @@ export default {
   randomFillSync,
   pbkdf2,
   pbkdf2Sync,
+  privateDecrypt,
+  privateEncrypt,
+  publicDecrypt,
+  publicEncrypt,
   randomBytes,
   randomUUID,
   scrypt,
@@ -33,6 +43,10 @@ export {
   Hash,
   pbkdf2,
   pbkdf2Sync,
+  privateDecrypt,
+  privateEncrypt,
+  publicDecrypt,
+  publicEncrypt,
   randomBytes,
   randomFill,
   randomFillSync,